requirements.txt
requests~=2.27.1
pandas~=1.4.2
psycopg2~=2.9.3
PyYAML~=6.0